server/pyproject.toml
[tool.poetry]
name = "baking"
version = "0.1.1"
description = ""
authors = ["Moshe Reubinoff <reubinoff@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10.0"

colorlog = "^6.4.1"
pydantic =  {extras = ["dotenv"], version = "^1.10.2"}
uvicorn = "0.20.0"
azure-storage-blob = "^12.13.1"
cryptography = "^39.0.0"
fastapi = "^0.95.0"
pandas = "^2.0.0"
motor = "^3.1.1"



[tool.poetry.dev-dependencies]
pylint = "^2.11.1"
mypy = "^0.910"
black = "^23.1.0"
autopep8 = "^1.5.7"
factory-boy = "^3.2.0"
requests = "^2.26.0"
pytest-cov = "^3.0.0"
hypothesis = "^6.23.3"
schemathesis = "^3.10.1"
pytest = "^7.2.2"
nox = "^2022.11.21"
nox-poetry = "^1.0.2"
flake8 = "^6.0.0"
pytest-asyncio = "^0.21.0"
polyfactory = "^2.0.0"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[build-system]
requires = ["poetry-core>=1.2.*"]
build-backend = "poetry.core.masonry.api"
